Client: PSM tracks stop-event
Previously, the stop event was handled by delegating it to pause().
That’s because our initial implementation was only meant to work with
spotify and spotify does not know a stop state.

This patch extends the PSM to handle stop events specifically. If a stop
event occurs, PSM does not care about the track but merely decides if it
is completed or cancelled and subsequently clears the state so that PSM
is reset completely.