rec: Extend mbs client to use new identify API
Extend the internal mbs client to be able to use the mbs v2 API, which
offers a more reliable method to identify playbacks. The mbs client is
simply extended by a new method, since we want to get rid of mbs v1 as
soon as possible. There is no need to create a new client for the new
API.