rec: Add updateMbsOnPlaybackGroup() to fix playbacks
The internal API is extended by updateMbsOnPlaybackGroup(). The method
can be used to set rgId and recId of a "playback group". Such a group is
a set of playbacks with identical original artist, release, recording
and length (including NULL of the latter).

This differs from fixAccumulatedPlaybacks() in that it does not relate
to any user. Consequently

1) we do not require this to be executed within a request context, we do
   not need any user.
2) it is not meant to be exposed on the API but merely an internal
   method.

This method will be used to repair unattached ("broken") playbacks by
looking them up via mbs.