rec: Extend getUnattachedPlaybackGroups() by recency filter
When loading unattached playback groups, the caller must specify the
allowed recency of any group member’s last fix attempt. That is, the
caller must specify a point in time t, such that only groups are
returned in which the most recent fix attempt on any member is before t.

This can be used to reduce load for candidates that will likely not be
fixed, because the most likely reason for a failed fix is a missing
entry in musicbrainz.