rec: Replace PlaybackJpaRepo by PlaybackRepo without Hibernate
A bit of a larger patch that removes the PlaybackBoundary’s dependency
on PlaybackJpaRepo. The logic is provided by a hibernate-free
PlaybackRepo. Some minor refactorings are entangled with those changes
and also contained in this patch (or due to my laziness).

We use our own sequence to define new oids for inserted playbacks, they
start at the highest oid + 1 from the current playback table.