rec: Provide "fixlog" endpoint to make autofixing traceable
Add endpoint .../playbacks/fixlog, which can be used to read a paged
fix log. The log’s entries are playbacks that are attached and have a
fix attempt (i.e. there was a successful fix attempt which implies that
they were unattached before and at the time of "fix attempt" were
repaired). These playbacks are ordered by the fix time and thus the list
can be read as a log.

A log for (failed) attempts on unattached playbacks may be added later.