client: Add experimental 'recs' and 'rgs' commands for simple stats
Add a new StatsService and StatsClient that will be used to retrieve
stats from the backend (or potentially calculate local stats with the
service layer later).

Add two commands 'recs' for top recordings and 'rgs' for top release
groups.

Then there is some intermingled refactoring going on, that I cannot be
arsed to pull into a separate commit:
* The formatter gets a new show() sub, that figures out (using the
  blessed class name) which printing sub shall be used. In this commit
  it only supports the newly added sub, but the others will follow. The
  nice thing: we have to implement this logic only in the Formatter role
  (which acts as our API to callers) and all formatters just have to
  implement the specific printing methods.
* Reorder the command mapping in Cli.pm
* Introduce required subs for the Formatter role. This is actually nice,
  because it helps remembering that all formatters should implement the
  required subs.