client: Make formatter a bean
We used to inject the formatter with a hack overriding the instance()
sub of Singleton. Turns out this sort of disturbed the mechanics of
Moo::Role, in that subs from this role were not visible in subtypes.
Since it was always a hack, we now implement the behaviour differently.

Remove the crappy hack code from the Formatter role.

Create a new bean "formatter" that does the choosing logic, depending on
the config value "format". Note, that this features the same lazy/late
init behaviour. Since the formatter is sort of special, it shouldn’t be
to bad to create a bean for it.

Adapt the call site (wich is merely replacing the singleton injection by
a inject call).