client: Load modules on demand only
This patch changes the bevhaviour of 'autowire' and 'singleton' (in a
backward compatible manner). The commands now load the respective module
before requiring it.

This means that we load a lot less modules for most invocations! Of
course it is a difficult question whether this speeds up or slows down,
but for cfm commands it reduces execution time by about 100ms wallclock
time.