client: Pull list command into dedicated package and document it
Add a --help or -h flag to the config.

Introduce a new command mapping in Cli.pm. This command mapping maps to
package names instead of subs. We can pull out command by command into
its own package. While there are "new" commands (in dedicated packages)
and old ones (simple subs in Cli.pm) simultaneously, we check both
mappings, where the old one has precedence. Once all commands are moved,
we remove the old mapping.

Add a Command role that expects subs run() and help(). run() must be
implemented by the respective command. help() is predefined to use
pod2usage() to print the documentation in that command as help, but it
can be overridden of course. Both commands receive the "subcommand
arguments", so we can implement different behaviour.

Pull out the "list" command to its own package and provide documentation
for it.

Remove the test for the list command: the way we test is not really
sustainable, we should add proper "integration" tests that can deal with
all the DI crap and so on later.