client: Introduce kv table
This is part of a sequence of patches to maintain local state on the
client. We will offer means to sync playbacks (and possibly other user
data) to the client. To control the sync state, we require some metadata
to be stored on the client.

As we will keep the synced user data in sqlite, it is an obvoius
solution to keep the sync meta data there as well.

Hence, this patch introduces a general purpose key-value mapping. It
maps strings to strings. Along with the internal interface, we introduce
some "internal" commands, namely kv, kv-set and kv-unset to examine,
update or delete the kv table.